Newton's Game
The game of Newton is one where a player tries to be the first to get 5 in a row either vertically, horizontally or diagonally.  Two players alternate making moves.  On each turn they may enter the column number where their piece will be placed.  The pieces are inserted from the top and slides down as far as it can go in that column.  Enter 'r' to rotate a piece out of the bottom of a column to be dropped back in at the top of the column.  Enter 'x' to exit.

- Input can be upper or lower-case for 'r' and 'x'
- On each move you can either place a piece, or rotate a column.

Checking for a win should be done in this order:
        1. All rows, starting from the top
        2. All columns, starting from the left
        3. Upper-left to lower-right diagonals, starting from the upper-left corner
        4. Lower-left to upper-right diagonals, starting from the middle on the left side.

If a move results in a win for both players at once, the winning configuration that is found first (using the sequence described in the check-for-win step) is the winner.