Twisted Hangman
A program to play hangman where the program chooses the word and the user tries to guess the word.  The twist is that the program starts out with a dictionary of many words.  After each human guess, the program eliminates all words containing the letters guessed so far, making the word as hard to find as possible.  Correct user guesses only show up once the list has been narrowed down so far that all remaining words have the letter that the user guesses.

Each time the user guesses a letter the program removes from the words list all words containing that letter.  Eventually this no longer works, as eliminating all words with a particular letter means there are no words remaining.  At this point, the program finds the first word on the list and displays the user's guessed character in the position in which it is found in that word. At this point you must also eliminate any words that have duplicates of that letter in some additional position as compared to this first word on your word list. 

The program still tries to be tricky, however, so it will also keep any other words still on the list that also have that letter in that position. 
Giving input of '~' results in a "cheat" where the program displays all words that are still possibilities.
Once the program has to start displaying letters, it must display all occurrences of any guessed letter.  Note that this will affect which other words are still possibilities.