Balanced Expression Checker
A C program that will determine whether input is given with properly balanced symbols using a stack implemented as a dynamic array. For this program, the following symbol pairs are checked:
parentheses: ( )  
curly braces: { } 
square brackets: [ ] 
angle brackets: < >
If the program determines the expression is unbalanced, it will display which symbols are missing or expected.  It will also show where they were expected in order for the expression to have been balanced.